src: consolidate dependent-file updates for layout-fix batch
Completes the call-site/.cpp updates that accompany the high-priority
struct-layout fixes (WaveState/collision/entities/effects/particles/
Effect_Bada::Pass). The prior per-cluster commits staged headers; this
folds in the dependent sources the implementer agents also edited:

- collision: ColAABB/ColLine/ColSphere reference->accessor migration
  ripples in Bomb.cpp, Coin.cpp, Fruit.cpp, DebugFlags.cpp + Col*.cpp.
- entities/effects: SlashEntity.{h,cpp} binary offset map + port-only
  fields behind #ifndef __bada__; EffectImage m_VelOut/m_TransitionHash
  port-only members (fixes prior C2065); SlashModInfo::LoopingSound
  field order + float retypes in ItemInfo.{h,cpp}.
- particles: PSPParticle/PSPParticleEmitter trimmed to binary layout;
  manager-side parallel m_ParticleLists holds the port particle pools.
- FruitSaveData coin members moved to GameWork (ItemManager/PowerUp/
  Coin/PowerUpShop call sites); CoinChanceinator ODR dedup.

Builds clean (per implementer subagents; orchestrator cmake blocked by
auto-accept-edits approval race).

+++ src/game/ItemInfo.h
@@ -154,13 +154,9 @@ struct SlashSoundMods {
 // -----------------------------------------------------------------------
 struct LoopingSound {
     // +0x00  float    m_field0     runtime float; ctor 0.0f (vstr.32 s15 = 0.0f)
-    // DIFFERS: binary = float at +0x00; port names as m_SoundId int32_t for
-    //   readability; same 4-byte layout, different type interpretation.
-    int32_t  m_SoundId;
+    float    m_SoundId;
     // +0x04  float    m_field4     runtime float; ctor 0.0f (vstr.32 s15 = 0.0f)
-    // DIFFERS: binary = float at +0x04; port names as m_Phase int32_t for
-    //   readability; same 4-byte layout, different type interpretation.
-    int32_t  m_Phase;
+    float    m_Phase;
     // +0x08  char*    m_pLoopName  CloneString of XML "loop" attr; default NULL
     char*    m_pLoopName;
     // +0x0c  int32_t  m_State      runtime state; ctor 0, not parsed
